requirements.txt
amqp==1.4.6
ansicolors==1.0.2
anyjson==0.3.3
Babel==2.0
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.20
celery==3.1.18
certifi==2015.9.6.2
configparser==3.5.0b2
decorator==4.0.2
Django==1.8.3
django-celery==3.1.16
django-compat==1.0.8
django-cors-headers==1.1.0
django-crequest==2014.9.19
django-crispy-forms==1.4.0
django-extensions==1.5.6
django-grappelli==2.7.1
django-gulp==2.0.0
django-hijack==1.0.10
django-kombu==0.9.4
django-revproxy==0.9.0
django-tastypie==0.12.2
dos2unix==1
env-tools==2.0.0
flower==0.8.3
funcsigs==0.4
futures==2.2.0
influxdb==2.7.1
kombu==3.0.30
mock==1.3.0
nose==1.3.7
pbr==1.6.0
psutil==2.2.1
pyparsing==2.0.3
python-dateutil==2.4.2
python-mimeparse==0.1.4
pytz==2015.4
requests==2.7.0
six==1.10.0
tini==3.0.1
tornado==4.2.1
urllib3==1.10.1
wheel==0.24.0
xlrd==0.9.4
pyjwt>=0.0
django-jwt-auth==0.0.2
django-http-proxy==0.4.3

django-tinymce==2.0.1
django-supervisor==0.3.2
Unidecode==0.04.18
pillow==2.8.2
psycopg2==2.6.2

httplib2==0.9.2
pandas==0.17.1
numpy==1.11.1
scipy==0.18.0
sklearn>=0.0

ciso8601>=0.0
fabric==1.12.0
gunicorn==19.6.0
